semiauto and auto imports (#353)
