--- core/src/main/scala-3/cats/derived/all.scala
@@ -1,23 +1,0 @@
-package cats.derived
-
-object all extends
-  CommutativeMonoidDerivation,
-  CommutativeSemigroupDerivation,
-  ContravariantDerivation,
-  EmptyDerivation,
-  EmptyKDerivation,
-  EqDerivation,
-  FoldableDerivation,
-  FunctorDerivation,
-  HashDerivation,
-  InvariantDerivation,
-  MonoidDerivation,
-  MonoidKDerivation,
-  OrderDerivation,
-  PartialOrderDerivation,
-  ReducibleDerivation,
-  SemigroupDerivation,
-  SemigroupKDerivation,
-  ShowDerivation,
-  TraverseDerivation,
-  Instances
