+++ core/src/main/scala-3/cats/derived/emptyk.scala
@@ -3,8 +3,6 @@ package cats.derived
 import alleycats.{Empty, EmptyK}
 import shapeless3.deriving.{Const, K1}
 
-object emptyk extends EmptyKDerivation
-
 trait ProductEmptyK[T[x[_]] <: EmptyK[x], F[_]](
   using inst: K1.ProductInstances[T, F]
 ) extends EmptyK[F]:
@@ -14,6 +12,3 @@ trait EmptyKDerivation:
   extension (E: EmptyK.type)
     inline def derived[F[_]](using gen: K1.ProductGeneric[F]): EmptyK[F] =
       new ProductEmptyK[EmptyK, F]{}
-
-  given [X](using X: Empty[X]): EmptyK[Const[X]] with
-    def empty[A]: X = X.empty
