+++ core/src/main/scala-3/cats/derived/instances.scala
@@ -1,6 +1,7 @@
 package cats.derived
 
 import cats.*
+import alleycats.*
 import shapeless3.deriving.Const
 
 private[derived] trait Instances extends Instances1:
@@ -27,3 +28,6 @@ private[derived] trait Instances1:
   given [X](using X: Semigroup[X]): SemigroupK[Const[X]] with
     def combineK[A](x: Const[X][A], y: Const[X][A]): Const[X][A] =
       X.combine(x, y)
+
+  given [X](using X: Empty[X]): EmptyK[Const[X]] with
+    def empty[A]: X = X.empty
