+++ core/src/main/scala-3/cats/derived/package.scala
@@ -1,0 +1,52 @@
+package cats.derived
+
+import cats.*
+
+extension (F: Foldable.type)
+  inline def derived[F[_]]: Foldable[F] = DerivedFoldable[F]
+
+extension (F: Functor.type)
+  inline def derived[F[_]]: Functor[F] = DerivedFunctor[F]
+
+extension (F: Reducible.type)
+  inline def derived[F[_]]: Reducible[F] = DerivedReducible[F]
+
+extension (F: Traverse.type)
+  inline def derived[F[_]]: Traverse[F] = DerivedTraverse[F]
+
+object semiauto extends
+  CommutativeMonoidDerivation,
+  CommutativeSemigroupDerivation,
+  ContravariantDerivation,
+  EmptyDerivation,
+  EmptyKDerivation,
+  EqDerivation,
+  HashDerivation,
+  InvariantDerivation,
+  MonoidDerivation,
+  MonoidKDerivation,
+  OrderDerivation,
+  PartialOrderDerivation,
+  SemigroupDerivation,
+  SemigroupKDerivation,
+  ShowDerivation,
+  Instances:
+
+  inline def foldable[F[_]]: Foldable[F] = DerivedFoldable[F]
+
+  inline def functor[F[_]]: Functor[F] = DerivedFunctor[F]
+
+  inline def reducible[F[_]]: Reducible[F] = DerivedReducible[F]
+
+  inline def traverse[F[_]]: Traverse[F] = DerivedTraverse[F]
+
+
+object auto:
+
+  inline given [F[_]]: Foldable[F] = DerivedFoldable[F]
+
+  inline given [F[_]]: Functor[F] = DerivedFunctor[F]
+
+  inline given [F[_]]: Reducible[F] = DerivedReducible[F]
+
+  inline given [F[_]]: Traverse[F] = DerivedTraverse[F]
