+++ core/src/test/scala-3/cats/derived/CommutativeMonoidTests.scala
@@ -1,9 +1,9 @@
 package cats.derived
 
-import alleycats._
-import cats._
+import alleycats.*
+import cats.*
 import cats.kernel.CommutativeMonoid
-import cats.derived.all._
+import cats.derived.semiauto.*
 
 class CommutativeMonoidTests { //
   case class Foo(i: Int, b: Option[Int]) derives CommutativeMonoid
