+++ core/src/test/scala-3/cats/derived/EmptyTests.scala
@@ -1,8 +1,8 @@
 package cats.derived
 
-import alleycats._
-import cats._
-import cats.derived.all._
+import alleycats.*
+import cats.*
+import cats.derived.semiauto.*
 
 object EmptyTests:
   case class Foo(i: Int, b: IntTree) derives Empty
