+++ core/src/test/scala-3/cats/derived/MonoidKTests.scala
@@ -1,9 +1,8 @@
 package cats.derived
 
-import alleycats._
-import cats._
-import cats.derived.all._
-import cats.derived.all.given
+import alleycats.*
+import cats.*
+import cats.derived.semiauto.*
 
 class MonoidKTests { //
   case class Foo[A](i: String, l: List[A]) derives MonoidK
