+++ core/src/test/scala-3/cats/derived/OrderTests.scala
@@ -1,8 +1,8 @@
 package cats.derived
 
-import alleycats._
-import cats._
-import cats.derived.all._
+import alleycats.*
+import cats.*
+import cats.derived.semiauto.*
 
 class OrderTests { //
   case class Foo(i: Int, b: Option[String]) derives Order
