+++ core/src/test/scala-3/cats/derived/SemigroupKTests.scala
@@ -1,9 +1,9 @@
 package cats.derived
 
-import alleycats._
-import cats._
-import cats.derived.all._
-import cats.derived.all.given
+import alleycats.*
+import cats.*
+import cats.derived.semiauto.*
+import cats.derived.semiauto.given
 
 class SemigroupKTests { //
   case class Foo[A](i: String, l: List[A]) derives SemigroupK
