+++ core/src/test/scala-3/cats/derived/ShowTests.scala
@@ -1,8 +1,8 @@
 package cats.derived
 
-import alleycats._
-import cats._
-import cats.derived.all._
+import alleycats.*
+import cats.*
+import cats.derived.semiauto.*
 
 class ShowTests { //
   case class Foo(i: Int, b: Option[String]) derives Show
